requirements.txt
uvicorn
fastapi
typing
psycopg2
face-recognition
numpy
Pillow